breaking: Added flag for adding gitignore(s).
Users are now able to input the gitignores at all cases.That incl-
udes running the file with args or with the prompt.

Also a key change is that the `project_type` is the only input one
needs to enter to get started with the project creation. If any of
the existing flags are not provided values with, they are prompted
to input the values.

Added relevant help messages and revamped the ascii text.